requirements.txt
google-generativeai==0.3.1
python-dotenv==1.0.0
streamlit==1.25.0
fastapi==0.105.0
uvicorn==0.24.0
websockets==12.0
websocket-client==1.7.0
flask==2.2.3
requests==2.31.0 